+++ code/summing.c
@@ -1,9 +1,9 @@
 #include "sum.h"
 
 // Sums the N elements of the array v from left to right.
-double summingSeq(double* v, const int N)
+double summingSeq(double* v, const long long N)
 {
-	int i;
+	long long i;
 	double Sn = 0;
 	for(i=0; i<N; ++i)
 	{
@@ -15,9 +15,9 @@ double summingSeq(double* v, const int N)
 // Sums the N elements of the array v using OpenMP to sum different parts
 // of the array. The order of summation (and thus the exact result) is
 // non-deterministic.
-double summingShared(double* v, const int N)
+double summingShared(double* v, const long long N)
 {
-	int i;
+	long long i;
 	double Sn = 0;
 	#pragma omp parallel for schedule(static) reduction(+:Sn)
 	for(i=0; i<N; ++i)
@@ -31,9 +31,9 @@ double summingShared(double* v, const int N)
 // always sum the two smalles elements and store the result to the array.
 // The approach enhanches numerical stability at the cost of sorting.
 // NOTE: The array is wrecked.
-double summingSlow(double* v, const int N)
+double summingSlow(double* v, const long long N)
 {
-	int i, j;
+	long long i, j;
 	double tmp;
 	for(i=N-1; i>=1; --i)
 	{
